add a command line option to override `source` entry in `llmk.toml`
It would be convenient if `llmk` can typeset an alternative file keeping
configuration in `llmk.toml. `llmk` already supports passing a
positional argument to specify a typeset file, but it only interprets
magic comments in the specified file. This commit adds a command line
option to override `source` entry in `llmk.toml`.

The new option is restricted unless another new `-T`, `--temporary`
option is given. This limitation retains consistency of workflow
definition and avoid regular usage of the `--source` option.